add pool len

+++ pool.go
@@ -41,6 +41,6 @@ func (p *Pool) PutIndex(i uint) (ok bool) {
 }
 
 func (p *Pool) IsFree(i uint) (ok bool) { return p.freeBitmap.Get(i) }
-func (p *Pool) FreeLen() int            { return len(p.freeIndices) }
+func (p *Pool) FreeLen() uint           { return uint(len(p.freeIndices)) }
 func (p *Pool) MaxLen() uint            { return p.maxLen }
 func (p *Pool) SetMaxLen(x uint)        { p.maxLen = x }
